+++ src/lib/NavBar/NavBar.svelte
