+++ src/routes/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
     // @ts-ignore
-    import Navbar from "$lib/NavBar.svelte"
+    import Navbar from "$lib/NavBar/NavBar.svelte"
     import MainView from "$lib/MidView.svelte"
 
 </script>
